Add OnPremisesExtensionAttributes class and update Get-PnPMicrosoft365Group cmdlet to include extension attributes
