+++ CHANGELOG.md
@@ -15,7 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added `Standard` and `EDU_Staff` template support to `New-PnPTeamsTeam` and provision Teams templates through Microsoft Graph team templates. Note that `EDU_Class` and `EDU_PLC` are now also provisioned through Microsoft Graph team templates (`POST /teams`) instead of the previous Microsoft 365 Group creation and teamify flow. [#999](https://github.com/pnp/powershell/issues/999)
 - Added `-ExchangeApplicationPermissions`, `-ExchangeDelegatePermissions`, `-PowerBIApplicationPermissions`, `-PowerBIDelegatePermissions`, `-DataverseDelegatePermissions`, `-PowerAppsDelegatePermissions`, `-AzureServiceManagementDelegatePermissions` and `-ResourcePermissions` to `Register-PnPEntraIDApp`, the last one to request permissions on any other API. [#5424](https://github.com/pnp/powershell/pull/5424)
 - Refreshed the permissions shipping with the module, taking Microsoft Graph from 223 application and 294 delegated permissions to 705 and 795, and SharePoint from 9 and 18 to 20 and 29. [#5424](https://github.com/pnp/powershell/pull/5424)
-- Added `-IncludeExtensionAttributes` to `Get-PnPMicrosoft365Group`, which populates the `OnPremisesExtensionAttributes` property with the extension attributes 1-15 of a group. Microsoft Graph only populates these for groups that are synchronized from an on-premises Active Directory, not for cloud only groups. [#4091](https://github.com/pnp/powershell/issues/4091)
+- Added `-IncludeExtensionAttributes` to `Get-PnPMicrosoft365Group`, which populates the `OnPremisesExtensionAttributes` property with the extension attributes 1-15 of a group. Microsoft Graph only populates these for groups that are synchronized from an on-premises Active Directory, not for cloud only groups. [#5425](https://github.com/pnp/powershell/pull/5425)
 
 ### Changed
 - Changed `Register-PnPEntraIDApp` and `Register-PnPEntraIDAppForInteractiveLogin` to register exactly the requested permissions. An invocation whose permissions were all silently dropped, such as `-Scopes` holding only delegated permissions, previously fell back to the default permission set, so the resulting app can differ from before for the same command line. [#5424](https://github.com/pnp/powershell/pull/5424)
